Convert TickscriptEditor to TypesScript and add initial test
