+++ ui/mocks/kapacitor/apis/index.ts
@@ -1,0 +1,2 @@
+export const pingKapacitorVersion = jest.fn(() => Promise.resolve('2.0'))
+export const getLogStreamByRuleID = jest.fn(() => Promise.resolve())
