+++ ui/package.json
@@ -41,6 +41,7 @@
     "@types/prop-types": "^15.5.2",
     "@types/react": "^16.0.38",
     "@types/react-router": "3",
+    "@types/text-encoding": "^0.0.32",
     "autoprefixer": "^6.3.1",
     "babel-core": "^6.5.1",
     "babel-eslint": "6.1.2",
