+++ ui/src/types/index.ts
@@ -1,6 +1,6 @@
 import {AuthLinks, Organization, Role, User, Me} from './auth'
-import {AlertRule, Kapacitor} from './kapacitor'
 import {Query, QueryConfig, TimeRange} from './query'
+import {AlertRule, Kapacitor, Task} from './kapacitor'
 import {Source} from './sources'
 import {DropdownAction, DropdownItem} from './shared'
 
@@ -18,4 +18,5 @@ export {
   DropdownAction,
   DropdownItem,
   TimeRange,
+  Task,
 }
