+++ ui/tsconfig.json
@@ -8,7 +8,8 @@
       "react",
       "prop-types",
       "jest",
-      "react-router"
+      "react-router",
+      "text-encoding"
     ],
     "target": "es6",
     "module": "es2015",
